requirements.txt
GoogleAppEngineCloudStorageClient==1.9.5.0
NoseGAE==0.2.1
coverage==3.7.1
gaenv==0.1.7
google-api-python-client==1.2
httplib2
nose==1.3.1
oauth2client
protorpc-standalone==0.9.1
pycrypto
rednose==0.4.1
